Experiment: Integrated Manager
An addon for Integrated Dynamics to get some overview over your
networks. Currently only planned as a viewer. The code is also
very work in progress, unoptimized and uncleaned.

But it's enough to experience a small preview if you build this
yourself. (You'll need a github username and token with access
to github packages though.)

There are many things to be done before this gets to a releasable
state. Currently it's prettier than it is useful.